Crate core_crypto_macros

Attribute Macros

dispotent

Neologism to mean the opposite of idempotent. Methods of ‘core_crypto::Client’ marked with this have to insert/delete an entity in the keystore and change the number of entities persisted.

durable

Will drop current MLS group in memory and replace it with the one in the keystore. This simulates an application crash. Once restarted, everything has to be loaded from the keystore, memory is lost.

idempotent

!!! Not literally idempotent !!!

Derive Macros

Entity

Implements the Entity trait for the given struct. To be used internally inside the core-crypto-keystore crate only.