core_crypto_macros

Derive Macro Debug

#[derive(Debug)]
{
    // Attributes available to this derive:
    #[sensitive]
}

Implements the Debug trait for the given struct changing the formatting of Vec<u8> fields. Structs deriving this custom Debug debug Vec<u8> or Option<Vec<u8>> as a hex string. Fields can be marked as sensitive to obfuscate them when logged. This needs the field type to implement Obfuscate.