core_crypto_macros

Derive Macro Entity

#[derive(Entity)]
{
    // Attributes available to this derive:
    #[entity]
}

Implements the Entity trait and related traits for the given struct.

Intended to be used only within core-crypto-keystore.

All fields in an entity using this derive macro must have one of four types:

Vec<u8>

String

Option<Vec<u8>>

u32

There is one exception: the identity field can only be a Vec<u8> or String; the optional variant is forbidden in that context.

This is intentionally restrictive. If your entity has other field types, implement these traits manually.

Attributes

Attributes modify the code generated by this proc macro by adding annotations to the whole struct or to distinct fields. All attributes accepted by this proc macro are keyed with #[entity]. The attribute itself appears in the following parenthesis.

It is acceptable to combine multiple entity attributes together, simply by combining them within the parenthesis, separated by commas: #[entity(id, column_name="column")], for example.

#[entity(table_name = "my_table")]

This is an outer attribute which modifies the entire item. Its parameters:

table_name: Optional. When set, this specifies the SQL table name which will be used. Otherwise, one will be generated by transforming the item name to snake case and adding an s.

no_upsert: Optional flag. By default the generated save emits INSERT OR REPLACE. When this flag is present it emits a plain INSERT instead, and a uniqueness violation is mapped to CryptoKeystoreError::AlreadyExists.

#[entity(id)]

This is a field attribute which can be applied to at most 1 field in the item.

The field to which this attribute is applied must have type Vec<u8> or String. Otherwise, you must implement all relevant entity traits manually instead of relying on this macro.

If this attribute does not appear, the macro will search for a field whose name is id and assume that is an implicit id.

This attribute does not accept a value.

#[entity(column = "my_column")]

This is a field attribute which can be applied to any number of fields in the struct.

The column value will be used as the SQL column name for this field. If this is unset, the field’s name will be used as the sql column name.